refactor: SocialLoginUser 객체생성 로직을 UserInfoProvider로 이동
