+++ backend/src/test/java/com/darass/darass/auth/oauth/api/domain/UserInfoProviderTest.java
@@ -10,6 +10,7 @@ import com.darass.darass.auth.oauth.api.domain.dto.KaKaoAccount;
 import com.darass.darass.auth.oauth.api.domain.dto.Profile;
 import com.darass.darass.auth.oauth.api.domain.dto.SocialLoginResponse;
 import com.darass.darass.auth.oauth.exception.AuthenticationException;
+import com.darass.darass.user.domain.SocialLoginUser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
@@ -32,19 +33,20 @@ class UserInfoProviderTest {
     private ObjectMapper objectMapper;
 
     @Test
-    @DisplayName("카카오 api 서버에 액세스 토큰을 전송하여 유저 정보를 가져온다.")
-    void findSocialLoginResponse() throws JsonProcessingException {
-        SocialLoginResponse socialLoginResponse = new SocialLoginResponse(
-            "1", new KaKaoAccount("jujubat@kakao.com",
-            new Profile("우기"))
-        );
-        String expectedResult = objectMapper.writeValueAsString(socialLoginResponse);
+    @DisplayName("카카오 api 서버에 액세스 토큰을 전송하면 SocialLoginUser 객체를 리턴한다.")
+    void findSocialLoginUser() throws JsonProcessingException {
+        Profile profile = new Profile("우기");
+        KaKaoAccount kaKaoAccount = new KaKaoAccount("jujubat@kakao.com", profile);
+        SocialLoginResponse socialLoginResponse = new SocialLoginResponse("1", kaKaoAccount);
 
+        String expectedResult = objectMapper.writeValueAsString(socialLoginResponse);
         mockServer.expect(requestTo(UserInfoProvider.KAKAO_API_SERVER_URI))
             .andRespond(withSuccess(expectedResult, MediaType.APPLICATION_JSON));
 
-        SocialLoginResponse actual = userInfoProvider.findSocialLoginResponse("test");
-        assertThat(actual.getId()).isEqualTo(socialLoginResponse.getId());
+        SocialLoginUser socialLoginUser = userInfoProvider.findSocialLoginUser("test");
+
+        assertThat(socialLoginUser.getNickName()).isEqualTo(profile.getNickname());
+        assertThat(socialLoginUser.getEmail()).isEqualTo(kaKaoAccount.getEmail());
     }
 
     @Test
@@ -53,7 +55,7 @@ class UserInfoProviderTest {
         mockServer.expect(requestTo(UserInfoProvider.KAKAO_API_SERVER_URI))
             .andRespond(withUnauthorizedRequest());
 
-        assertThatThrownBy(() -> userInfoProvider.findSocialLoginResponse("test"))
+        assertThatThrownBy(() -> userInfoProvider.findSocialLoginUser("test"))
             .isInstanceOf(AuthenticationException.class)
             .hasMessage("토큰 인증에 실패하였습니다.");
     }
